Simplified and refactored code
Refactored deck_of_cards.c from five functions to two and made it more flexible. Got rid of using an integer variable array as a pointer to the card since we can just move the struct itself around in its own array.

Refactored blackjack_test as well to be more modular as well. It's now much shorter with just as much if not more functionality.