refactor: Use score in Hand struct
Change deal_card() to update the score member of the Hand struct every time it is called. Also update the logging to show the update count in addition to the card dealt.

Replace all occurences of calls to blackjack_score() (and variables storing the return result) with references to the correct hand.score in check_dealer_hand(), play_hands(), play_dealer_hand() and check_table().